Move PermissionGroupForm to a forms module
